o  updated XML-RPC for int Ids.
git-svn-id: https://svn.apache.org/repos/asf/maven/continuum/branches/id-refactor@497972 13f79535-47bb-0310-9956-ffa450edef68

+++ continuum-xmlrpc/src/main/java/org/apache/maven/continuum/xmlrpc/ContinuumXmlRpc.java
@@ -39,9 +39,9 @@ public interface ContinuumXmlRpc
     // Projects
     // ----------------------------------------------------------------------
 
-    Hashtable removeProject( int projectId );
+    Hashtable removeProject( long projectId );
 
-    Hashtable getProject( int projectId );
+    Hashtable getProject( long projectId );
 
     Hashtable getProjects();
 
@@ -49,7 +49,7 @@ public interface ContinuumXmlRpc
     // Building
     // ----------------------------------------------------------------------
 
-    Hashtable buildProject( int projectId, int trigger );
+    Hashtable buildProject( long projectId, int trigger );
 
     // ----------------------------------------------------------------------
     // Builds
